@@ -28,7 +28,7 @@ func jsonParam() usecase.Interactor {
 		JSONPayload
 	}
 
-	u := usecase.NewIOI(new(inputWithJSON), new(outputWithJSON), func(ctx context.Context, input, output interface{}) (err error) {
+	u := usecase.NewIOI(new(inputWithJSON), new(outputWithJSON), func(ctx context.Context, input, output any) (err error) {
 		var (
 			in  = input.(*inputWithJSON)
 			out = output.(*outputWithJSON)
