@@ -45,8 +45,8 @@ func (re RequestErrors) Error() string {
 }
 
 // Fields returns request errors by field location and name.
-func (re RequestErrors) Fields() map[string]interface{} {
-	res := make(map[string]interface{}, len(re))
+func (re RequestErrors) Fields() map[string]any {
+	res := make(map[string]any, len(re))
 
 	for k, v := range re {
 		res[k] = v
