@@ -30,7 +30,7 @@ func Test_decodeJSONBody(t *testing.T) {
 	assert.Equal(t, "248df4b7-aa70-47b8-a036-33ac447e668d", i.CustomerID)
 	assert.Equal(t, "withdraw", i.Type)
 
-	vl := rest.ValidatorFunc(func(in rest.ParamIn, namedData map[string]interface{}) error {
+	vl := rest.ValidatorFunc(func(in rest.ParamIn, namedData map[string]any) error {
 		return nil
 	})
 
@@ -90,7 +90,7 @@ func Test_decodeJSONBody_validateFailed(t *testing.T) {
 
 	var i []int
 
-	vl := rest.ValidatorFunc(func(in rest.ParamIn, namedData map[string]interface{}) error {
+	vl := rest.ValidatorFunc(func(in rest.ParamIn, namedData map[string]any) error {
 		return errors.New("failed")
 	})
 
