@@ -5,7 +5,7 @@ import "encoding/json"
 // Validator validates a map of decoded data.
 type Validator interface {
 	// ValidateData validates decoded request/response data and returns error in case of invalid data.
-	ValidateData(in ParamIn, namedData map[string]interface{}) error
+	ValidateData(in ParamIn, namedData map[string]any) error
 
 	// ValidateJSONBody validates JSON encoded body and returns error in case of invalid data.
 	ValidateJSONBody(jsonBody []byte) error
@@ -15,10 +15,10 @@ type Validator interface {
 }
 
 // ValidatorFunc implements Validator with a func.
-type ValidatorFunc func(in ParamIn, namedData map[string]interface{}) error
+type ValidatorFunc func(in ParamIn, namedData map[string]any) error
 
 // ValidateData implements Validator.
-func (v ValidatorFunc) ValidateData(in ParamIn, namedData map[string]interface{}) error {
+func (v ValidatorFunc) ValidateData(in ParamIn, namedData map[string]any) error {
 	return v(in, namedData)
 }
 
@@ -29,14 +29,14 @@ func (v ValidatorFunc) HasConstraints(_ ParamIn) bool {
 
 // ValidateJSONBody implements Validator.
 func (v ValidatorFunc) ValidateJSONBody(body []byte) error {
-	return v(ParamInBody, map[string]interface{}{"body": json.RawMessage(body)})
+	return v(ParamInBody, map[string]any{"body": json.RawMessage(body)})
 }
 
 // RequestJSONSchemaProvider provides request JSON Schemas.
 type RequestJSONSchemaProvider interface {
 	ProvideRequestJSONSchemas(
 		method string,
-		input interface{},
+		input any,
 		mapping RequestMapping,
 		validator JSONSchemaValidator,
 	) error
@@ -47,7 +47,7 @@ type ResponseJSONSchemaProvider interface {
 	ProvideResponseJSONSchemas(
 		statusCode int,
 		contentType string,
-		output interface{},
+		output any,
 		headerMapping map[string]string,
 		validator JSONSchemaValidator,
 	) error
@@ -63,7 +63,7 @@ type JSONSchemaValidator interface {
 
 // RequestValidatorFactory creates request validator for particular structured Go input value.
 type RequestValidatorFactory interface {
-	MakeRequestValidator(method string, input interface{}, mapping RequestMapping) Validator
+	MakeRequestValidator(method string, input any, mapping RequestMapping) Validator
 }
 
 // ResponseValidatorFactory creates response validator for particular structured Go output value.
@@ -71,7 +71,7 @@ type ResponseValidatorFactory interface {
 	MakeResponseValidator(
 		statusCode int,
 		contentType string,
-		output interface{},
+		output any,
 		headerMapping map[string]string,
 	) Validator
 }
@@ -87,8 +87,8 @@ func (re ValidationErrors) Error() string {
 }
 
 // Fields returns request errors by field location and name.
-func (re ValidationErrors) Fields() map[string]interface{} {
-	res := make(map[string]interface{}, len(re))
+func (re ValidationErrors) Fields() map[string]any {
+	res := make(map[string]any, len(re))
 
 	for k, v := range re {
 		res[k] = v
